requirements.txt
aiohttp
redis
pillow==9.5.0
discord.py
python-dotenv
easy-pil
pymongo
lavalink==5.3.0
groq==0.5.0
better_profanity
pickledb
cryptography
deep-translator
discord-interactions.py
discord.ui
gtts
asyncpraw
fastapi
uvicorn
aioredis
